fix: can now compile on linux

+++ include/pcap_constants.h
@@ -6,6 +6,7 @@
 #define PCAP_CONSTANTS_H
 
 #include <cstdint>
+#include <cstddef>
 
 namespace pcap_constants {
     // PCAP файл константы
@@ -34,7 +35,7 @@ namespace pcap_constants {
     static const size_t IPV6_DST_ADDR_OFFSET = 24;
 
     // Бенчмарки
-    static const size_t TRIALS = 50;
+    static const int TRIALS = 50;
 } // namespace pcap_constants
 
 #endif //PCAP_CONSTANTS_H
